src/third_party/blink/web_tests/svg/custom/feDisplacementMap-invalid-selector-expected.txt

CONSOLE ERROR: line 6: Error: <feDisplacementMap> attribute yChannelSelector: Unexpected end of attribute. Unrecognized enumerated value, "".
